fix: resolve remaining 5 golangci-lint issues
errcheck (3): add _ = f.sock.Close() in error paths of UDPFramer New
unconvert (1): remove unnecessary uint16(datalen) in l2slave write path
gofmt (1): format udp_test.go

All tests pass, go vet clean

+++ internal/sim/l2slave.go
@@ -337,7 +337,7 @@ func (s *L2Slave) ProcessFrame(infr *ecfr.Frame) (ofr *ecfr.Frame) {
 				copy(dg.Data[:datalen], s.BackingMemory[physbase:physbase+datalen])
 			}
 			if writes {
-				copy(s.BackingMemory[physbase:physbase+uint16(datalen)], dg.Data[:datalen])
+				copy(s.BackingMemory[physbase:physbase+datalen], dg.Data[:datalen])
 			}
 			// WKC: always increment for non-register memory (no masking)
 			if reads && writes {
